Update cabal file
One nice thing is that beginning from Cabal 2.4 we can use ** wildcard in
paths, which is especially useful in our case because we have many examples
in a big directory tree, so without ** listing all the directories correctly
is a real pain.

All other changes are just to satisfy Cabal's nitpicks.